add abstract command tracker
